+++ NEWS.md
@@ -24,6 +24,7 @@
 ## Bugfixes
 
 -   `seasonally()` result will be on the 15th of the centre month of each season instead of on the 1st. This makes the date more representative of the time span and also solves a bug in which dates on the 31st would return `NA`. This is a **potentially breaking change**.
+-   `ReadNetCDF()` doesn't fail when reading variables with no dimensions (thanks to @paocorrales, \#141).
 
 # metR 0.9.2